Completion of registration + login + session creation
Comments/changes to files
conn.php: commented out the print statement after succesful connection to DB

verifyloggedinstatus.php + myreviews.php: To check if there is a logged on user. If user is logged on, it would print username and user_id. If not logged in, would redirect to login page.

login.php + registration.php: Redirects user to myreviews.php is there is already a login session

ict2206-db.sql: Contains 1 table which contains the user_id, username and password column